Force reauthentication after 24 hours (#50)
* Force reauthentication after 24 hours

This is enforced with two minor changes together:
- Set the default session length within django to 24 hours
- Use the 'prompt' argument when sending user to log in with Auth0.
This means that even if the user is still logged in to Auth0 when
the django session expires, they need to reenter their credentials
again to access the Dashboard Service. Without using the 'prompt'
argument, the user would be silently reauthenticated by auth0 until
their session has expired.

An alternative to using 'prompt' would be to change the session length
setting within Auth0. However, this change would affect all apps in
the tenant, not just Dashboard Service.