Feature/entra (#76)
* Remove login prompt

To avoid users logged in with Entra having to repeat the process

* Add new fields to User model

* Override username, save and add custom UserManager

There is no need to store the username, as we will use the email
field instead. So this commit drops the username field and adds a
custom UserManager to handle user creation.
Also override the save method to ensure that email is always stored
lowercase.

* Update logic to create or update users

Change code to lookup users based on their Auth0 ID.
Fallback to lookup users based on their email.
This will result in us updating existing users to store the
auth0 and entra ID values. We will then use these to lookup users
in future, and update their email if this changes. This makes the
auth process more robust, rather than relying on username/email
values taht could change.

* Update unit tests

* Add unit tests for new auth methods

* Fix test patching error

* Bump chart version to 0.2.0