Dont update auth0_id unless it is from Azure (#78)
* Dont update auth0_id unless it is from Azure

As the purpose of this work is to help move away from Auth0, towards Azure/Entra ID, where possible we should
be capturing the user ID from Entra. Previously there was an edge case where this would be overridden if a
user was able to log in to the app via the code method, which could occur if they were already logged in to
an app via the code method. This PR fixes that edge case by only updating the auth0_id if we dont already
have one, or if the provider is the Azure Entra ID connection

* Bump chart version